Update 'flutter pub' run to 'dart run'
The flutter pub run command is no longer supported in Flutter 2.10.0 and later. You should use the dart run command instead. The flutter pub run command was used to run Dart scripts that were published to pub.dev. However, this command was not always reliable and could sometimes cause problems. The dart run command is a more reliable way to run Dart scripts, and it also supports a wider range of features.