Update README.md
Changed "flutter_launcher_icons" to "flutter_icons" to fix broken example. 
The dependency should be "flutter_launcher_icons" (as is) but the paths need to be specified under "flutter_icons" (currently the example says "flutter_launcher_icons" which results in this error:

✗ ERROR: NoConfigFoundException
No configuration found in flutter_launcher_icons.yaml or in pubspec.yaml. In case file exists in different directory use --file option
#0      createIconsFromArguments (package:flutter_launcher_icons/main.dart:85:7)


but changing it to "flutter_icons" works: 
Creating Icons for Web...
Creating Icons for Windows...
Creating Icons for MacOS...

✓ Successfully generated launcher icons